Extract timeline overlay styles
